+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Update dependencies
   - New APIs:
     - GetBucketV2 (ListObjectsV2): `Aliyun.Oss.Bucket.list_objects`
+    - GetObjectMeta: `Aliyun.Oss.Object.get_object_meta`
   - **Breaking changes**:
     - Use separated modules to scope SubResources of Bucket/Object, for example:
       - `Aliyun.Oss.Bucket.get_acl` becomes `Aliyun.Oss.Bucket.ACL.get`